chore: A bit of reorganizing
